# frontend/lib/api/generated/component1-openapi.json
{"openapi":"3.1.0","info":{"title":"Component 1 - Liquidity-Aware Forecasting & Portfolio Optimization","description":"Liquidity-aware forecasting and multi-objective portfolio optimization with instant, loss-minimized withdrawal planning (J26-SE-325, Component 1).","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","description":"Liveness plus the active model version and which forecasters loaded.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/forecast":{"post":{"summary":"Forecast","description":"Per-symbol quantile forecast from the active hybrid model.","operationId":"forecast_forecast_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForecastRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForecastResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/portfolio/optimize":{"post":{"summary":"Optimize Portfolio","description":"Long-term allocation via MOEA/D. Publishes a decision event.","operationId":"optimize_portfolio_portfolio_optimize_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PortfolioRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PortfolioResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/portfolio/withdraw":{"post":{"summary":"Withdraw","description":"Instant, loss-minimized withdrawal plan via the fuzzy GA. Publishes a decision event.\n\nWith use_agent=True the Phase 5c agent chooses the optimizer inputs; the numbers still\ncome from the GA either way.\n\nAn infeasible plan is returned with HTTP 200 and `feasible=false`, not an error status.\n\"You can raise 80k of the 300k you asked for, and here is the cheapest way to do it\" is\na successful, actionable answer -- and RQ4 depends on infeasible plans being observable\nrather than thrown away.","operationId":"withdraw_portfolio_withdraw_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"AssetSale":{"properties":{"symbol":{"type":"string","title":"Symbol"},"sell_fraction":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Sell Fraction"},"quantity":{"type":"number","minimum":0.0,"title":"Quantity"},"expected_price":{"type":"number","title":"Expected Price"},"expected_slippage_pct":{"type":"number","title":"Expected Slippage Pct"},"execution_day":{"type":"integer","title":"Execution Day"}},"type":"object","required":["symbol","sell_fraction","quantity","expected_price","expected_slippage_pct","execution_day"],"title":"AssetSale","description":"One line of the liquidation plan. Every number here originates in the fuzzy GA."},"ForecastRequest":{"properties":{"symbols":{"items":{"type":"string"},"type":"array","title":"Symbols"},"horizon":{"type":"integer","maximum":21.0,"minimum":1.0,"title":"Horizon","default":5}},"type":"object","required":["symbols"],"title":"ForecastRequest"},"ForecastResponse":{"properties":{"forecasts":{"items":{"$ref":"#/components/schemas/SymbolForecast"},"type":"array","title":"Forecasts"},"model_version":{"type":"string","title":"Model Version"},"generated_at":{"type":"string","format":"date-time","title":"Generated At"}},"type":"object","required":["forecasts","model_version","generated_at"],"title":"ForecastResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"model_version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model Version"},"kafka_connected":{"type":"boolean","title":"Kafka Connected"},"available_forecasters":{"items":{"type":"string"},"type":"array","title":"Available Forecasters"}},"type":"object","required":["status","model_version","kafka_connected","available_forecasters"],"title":"HealthResponse"},"Holding":{"properties":{"symbol":{"type":"string","title":"Symbol"},"quantity":{"type":"number","minimum":0.0,"title":"Quantity"},"current_price":{"type":"number","exclusiveMinimum":0.0,"title":"Current Price"},"avg_daily_volume":{"type":"number","minimum":0.0,"title":"Avg Daily Volume"},"cost_basis":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Cost Basis"}},"type":"object","required":["symbol","quantity","current_price","avg_daily_volume"],"title":"Holding","description":"One position in the simulated portfolio."},"PortfolioRequest":{"properties":{"holdings":{"items":{"$ref":"#/components/schemas/Holding"},"type":"array","title":"Holdings"},"risk_preference":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Risk Preference","default":0.5},"selection_rule":{"type":"string","title":"Selection Rule","default":"knee"},"max_weight":{"type":"number","maximum":1.0,"exclusiveMinimum":0.0,"title":"Max Weight","default":0.25},"allow_shorting":{"type":"boolean","title":"Allow Shorting","default":false}},"type":"object","required":["holdings"],"title":"PortfolioRequest"},"PortfolioResponse":{"properties":{"recommended_weights":{"additionalProperties":{"type":"number"},"type":"object","title":"Recommended Weights"},"expected_return":{"type":"number","title":"Expected Return"},"expected_cvar":{"type":"number","title":"Expected Cvar"},"expected_liquidity_cost":{"type":"number","title":"Expected Liquidity Cost"},"pareto_front_size":{"type":"integer","title":"Pareto Front Size"},"selection_rule":{"type":"string","title":"Selection Rule"},"selection_rationale":{"type":"string","title":"Selection Rationale"},"model_version":{"type":"string","title":"Model Version"},"generated_at":{"type":"string","format":"date-time","title":"Generated At"}},"type":"object","required":["recommended_weights","expected_return","expected_cvar","expected_liquidity_cost","pareto_front_size","selection_rule","selection_rationale","model_version","generated_at"],"title":"PortfolioResponse"},"SymbolForecast":{"properties":{"symbol":{"type":"string","title":"Symbol"},"horizon":{"type":"integer","title":"Horizon"},"p10":{"type":"number","title":"P10"},"p50":{"type":"number","title":"P50"},"p90":{"type":"number","title":"P90"}},"type":"object","required":["symbol","horizon","p10","p50","p90"],"title":"SymbolForecast"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WithdrawalRequest":{"properties":{"holdings":{"items":{"$ref":"#/components/schemas/Holding"},"type":"array","title":"Holdings"},"target_amount":{"type":"number","exclusiveMinimum":0.0,"title":"Target Amount"},"urgency":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Urgency","default":0.5},"risk_tolerance":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Risk Tolerance","default":0.5},"deadline_days":{"type":"integer","maximum":30.0,"minimum":1.0,"title":"Deadline Days","default":1},"use_agent":{"type":"boolean","title":"Use Agent","default":false}},"type":"object","required":["holdings","target_amount"],"title":"WithdrawalRequest"},"WithdrawalResponse":{"properties":{"assets_to_sell":{"items":{"$ref":"#/components/schemas/AssetSale"},"type":"array","title":"Assets To Sell"},"raised_amount":{"type":"number","title":"Raised Amount"},"target_amount":{"type":"number","title":"Target Amount"},"shortfall":{"type":"number","title":"Shortfall"},"expected_slippage_pct":{"type":"number","title":"Expected Slippage Pct"},"expected_realized_loss":{"type":"number","title":"Expected Realized Loss"},"residual_portfolio_weights":{"additionalProperties":{"type":"number"},"type":"object","title":"Residual Portfolio Weights"},"days_required":{"type":"integer","title":"Days Required"},"feasible":{"type":"boolean","title":"Feasible"},"model_version":{"type":"string","title":"Model Version"},"fuzzy_rule_trace":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Fuzzy Rule Trace"},"agent_reasoning_trace":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Agent Reasoning Trace"},"generated_at":{"type":"string","format":"date-time","title":"Generated At"}},"type":"object","required":["assets_to_sell","raised_amount","target_amount","shortfall","expected_slippage_pct","expected_realized_loss","residual_portfolio_weights","days_required","feasible","model_version","generated_at"],"title":"WithdrawalResponse","description":"The instant-withdrawal plan -- this component's headline output."}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}